Fix the index comment fetching : Don't purge the comments of an index before fetching them else you can't know if a comment is new or not ...
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/apps/Thaw@14709 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/thaw/plugins/index/Index.java
@@ -1027,7 +1027,7 @@ public class Index extends Observable implements MutableTreeNode,
 			if (c.exists()) {
 				nmbFailedCommentFetching = 0;
 
-				if (c.isNew()) {
+				if (c.isNew() && c.isValid()) {
 					Logger.info(this, "New comment !");
 
 					setNewCommentFlag(true);
@@ -1081,7 +1081,8 @@ public class Index extends Observable implements MutableTreeNode,
 	public void purgeIndex(boolean useDontDelete) {
 		purgeFileList(useDontDelete);
 		purgeLinkList(useDontDelete);
-		purgeCommentKeys();
+		if (!useDontDelete)
+			purgeCommentKeys();
 	}
 
 
@@ -1447,6 +1448,12 @@ public class Index extends Observable implements MutableTreeNode,
 	}
 
 	public boolean setNewCommentFlagInMem(boolean flag) {
+		try {
+			throw new Exception("bleh");
+		} catch(Exception e) {
+			e.printStackTrace();
+		}
+
 		newComment = flag;
 		return true;
 	}
